Add files via upload
This is my project for the Advanced Database and Data Mining to Professor Joseph C. Lorilla as compliance for the 2nd Semester subject.

About the Project
--------------------

This project requires data sets to be interpreted or anlyzed using the Python as reference. In this case, I use the data sets from Kaggle particularly dabase of Datafiniti's Business as to their Hotel Reviews through state-by-state basis. However, I made some changes by deleting some of non-english reviews so that it will not be conflicted to Wordcloud and also in consideration to the data sets needed for the project.

Process/Method Used
-------------------------
This analysis use Sentiment Analysis, WordCloud,  MultinomialNaiveBayes and Confusion Matrix as basis of interpretation to the data sets. This process also obtain research and survey as to the use of Hotel Reviews of 1000 Hotels in a state-by-state basis and the following method was used:
NLTK: the most famous python module for NLP techniques
Numpy: offers comprehensive mathematical functions, random number generators, linear algebra routines, Fourier transforms, and moret
Scikit-learn: the most used python machine learning library
TensorFlow: is an end-to-end open source platform for machine learning.
Panda: is used to analyze data
Word Cloud: is a technique to show which words are the most frequent among the given text.
Seaborn: data visualization library based on matplotlib
Matplotlib: is a comprehensive library for creating static, animated, and interactive visualizations
Collections:  implements specialized container datatypes providing alternatives to Python's general purpose built-in containers, dict , list , set , and tuple .
BS4: library for pulling data out of HTML and XML files.

Expected Output
--------------------

1. Calculate the average accuracy as to the accuracy test score
2. Presented the most frequent words used through an image with text sizes emphasizing words with common use on the review
3. Presented the graph of the sentiment analysis